requirements.txt
# requirements.txt
# Install everything with:  pip install -r requirements.txt

streamlit>=1.35.0
groq>=0.9.0
requests>=2.31.0
google-generativeai>=0.7.0
cohere>=5.0.0
python-dotenv>=1.0.0

# Ollama uses the local REST API — no extra package needed (just requests)
# OpenAI is intentionally excluded from this project
